Feature: Implement anonymous nicknames and anonymous posts on categories page (#32)
* Rebase fix to only contain anonymous post updates

* feature(anonymous-posts): Implemnt anonymous posts
Changes:
- backend db flag in topics for their creation to show
user is anonymous
- backend db logic to handle anonymous posts to see
if the poster is the topic creater and set topic to anonymous
- backend logic added to posts to check if viewer is admin
or poster to make said post viewable instead of anonymous
- no frontend template changes were needed

To do:
- However fixes are required as avatar is still shown
for user and if clicked on in preview still goes to the user profile
- Also need to add the ability for user to edit their post to make it anonymous
- Also may want to add quality of life change to anonymous posts
pick random user name (e.g. anonymous beaver) when posting instead of just
Anonymous

* fix(tests): modify yaml files to reflect new schemas
Modified different yaml files to include
'anonymous' field in PostObject and TopicObject schemas.
to pass tests in npm run test.

* Rebase fix #2

* Rebase fix #3, removed extraneous generateAnonymousName func

* Fixed tab error in commment for src/api/topics.js

* public/src/client/topic/posts.js trailing comma lint fix

* fix(anonymous tests): update tests to reflect anonymous post changes
incorrect relative path in tests/anonymous.js for
anonymous post profile pic added relative path in tests
to mirror the actual path to the image